+++ stroom-app/src/main/java/stroom/app/App.java
@@ -128,7 +128,10 @@ public class App extends Application<Config> {
     public void run(final Config configuration, final Environment environment) {
         LOGGER.info("Using application configuration file {}", configFile.toAbsolutePath().normalize());
 
-        // Turn on Jersey logging.
+        // Turn on Jersey logging of request/response payloads
+        // I can't seem to get this to work unless Level is SEVERE
+        // TODO need to establish if there is a performance hit for using the JUL to SLF bridge
+        //   see http://www.slf4j.org/legacy.html#jul-to-slf4j
         environment.jersey().register(
                 new LoggingFeature(
                         java.util.logging.Logger.getLogger(LoggingFeature.DEFAULT_LOGGER_NAME),