+++ stroom-core/src/main/java/stroom/core/query/SuggestionsResourceImpl.java
@@ -17,6 +17,8 @@
 package stroom.core.query;
 
 import com.codahale.metrics.health.HealthCheck.Result;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import stroom.docref.DocRef;
 import stroom.meta.api.MetaService;
 import stroom.meta.shared.MetaFields;
@@ -32,14 +34,18 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Objects;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 class SuggestionsResourceImpl implements SuggestionsResource, HasHealthCheck {
+    private static final Logger LOGGER = LoggerFactory.getLogger(SuggestionsResourceImpl.class);
     private static final int LIMIT = 20;
 
     private final MetaService metaService;
     private final PipelineStore pipelineStore;
     private final SecurityContext securityContext;
+    private static final Pattern SEPARATOR_PATTERN = Pattern.compile("[_\\-]");
 
     @Inject
     SuggestionsResourceImpl(final MetaService metaService,
@@ -86,22 +92,86 @@ class SuggestionsResourceImpl implements SuggestionsResource, HasHealthCheck {
         });
     }
 
-    private List<String> createFeedList(final String text) {
+    private List<String> createFeedList(final String input) {
+        // TODO this seems pretty inefficient as each call hits the db to get ALL feeds
+        //   then limits/filters in java.  Needs to work off a cached feed name list
+
+        // TODO consider using isFuzzyMatch below
         return metaService.getFeeds()
                 .parallelStream()
-                .filter(name -> name == null || name.startsWith(text))
+                .filter(feedName -> feedName == null || input == null || feedName.startsWith(input))
                 .sorted(Comparator.naturalOrder())
                 .limit(LIMIT)
                 .collect(Collectors.toList());
     }
 
-    private List<String> createStreamTypeList(final String text) {
+    private List<String> createStreamTypeList(final String input) {
+        // TODO this seems pretty inefficient as each call hits the db to get ALL feeds
+        //   then limits/filters in java.  Needs to work off a cached feed name list
+
+        // TODO consider using isFuzzyMatch below
         return metaService.getTypes()
                 .parallelStream()
-                .filter(name -> name == null || name.startsWith(text))
+                .filter(typeName -> typeName == null || input == null || typeName.startsWith(input))
                 .sorted(Comparator.naturalOrder())
                 .limit(LIMIT)
                 .collect(Collectors.toList());
+    }
+
+    /**
+     * See tests for an idea of how this is meant to work. This is unfinished and
+     * needs optimising.
+     */
+    static boolean isFuzzyMatch(final String input, final String text) {
+        final String inputStr = Objects.requireNonNullElse(input, "");
+        final String testStr = Objects.requireNonNullElse(text, "");
+
+        LOGGER.debug("Testing input: {} against text: {}", inputStr, testStr);
+        if (testStr.startsWith(inputStr)) {
+            LOGGER.info("Matched on startsWith");
+            return true;
+        } else if (testStr.contains(inputStr)) {
+            LOGGER.info("Matched on contains");
+            return true;
+        } else {
+            boolean hasMatched = false;
+            String patternStr = inputStr.chars().boxed()
+                    .map(charVal -> {
+                        char theChar = (char) charVal.intValue();
+                        if (Character.isLetterOrDigit(theChar)) {
+                            return String.valueOf(theChar);
+                        } else if (SEPARATOR_PATTERN.matcher(String.valueOf(theChar)).matches()) {
+                            return "[^_-]*" + theChar;
+                        } else {
+                            throw new RuntimeException("Unexpected char " + theChar);
+                        }
+                    })
+                    .collect(Collectors.joining());
+            patternStr = ".*" + patternStr + ".*";
+
+            if (Pattern.matches(patternStr, testStr)) {
+                LOGGER.info("Matched on pattern: {}", patternStr);
+                hasMatched = true;
+            } else {
+                LOGGER.info("Not Matched on pattern: {}", patternStr);
+            }
+
+            patternStr = inputStr.chars().boxed()
+                    .map(charVal -> {
+                        char theChar = (char) charVal.intValue();
+                        return String.valueOf(theChar);
+                    })
+                    .collect(Collectors.joining(".*"));
+            patternStr = ".*" + patternStr + ".*";
+
+            if (Pattern.matches(patternStr, testStr)) {
+                LOGGER.info("Matched on pattern: {}", patternStr);
+                hasMatched = true;
+            } else {
+                LOGGER.info("Not Matched on pattern: {}", patternStr);
+            }
+            return hasMatched;
+        }
     }
 
     @Override